Balance credit production
* Adding production total to tax credits

In addition to a tax on the population and on 'bonuses' have added a tax on total production, in the ecomomy, by adding NetEnergy + NetFood + NetIndustry + NetIntelligence + NetResearch. Divided that by 3 to keep it modest for now. Need to better understand the whole economy to improve it more.

* Source of Production Total updated

NetEnergy, NetFood.... is not the total production capacity but a difference of production and consumption. Now using GetProductionOutput(ProductionCategory.Food) + ....

* Working form of Credit for testing

play test to see how well, or not, it works